Remove users/vpns from a group
- This feature is accessible by group admin/super admin only

+++ app/controllers/groups_controller.rb
@@ -8,15 +8,20 @@ class GroupsController < ApplicationController
   end
 
   def create
-    @group = Group.new(group_params)
-    respond_to do |format|
-      if @group.save
-        format.html { redirect_to groups_path, notice: 'Group was successfully created.' }
-        format.json { render status: :created, json: "#{@group.name}host created" }
-      else
-        format.html { redirect_to groups_path, notice: "Can't save '#{group_params[:name]}'" }
-        format.json { render status: :error, json: "#{@group.name} not created" }
+    if current_user.admin?
+      @group = Group.new(group_params)
+      respond_to do |format|
+        if @group.save
+          format.html { redirect_to groups_path, notice: 'Group was successfully created.' }
+          format.json { render status: :created, json: "#{@group.name}host created" }
+        else
+          format.html { redirect_to groups_path, notice: "Can't save '#{group_params[:name]}'" }
+          format.json { render status: :error, json: "#{@group.name} not created" }
+        end
       end
+    else
+      format.html { redirect_to groups_path, notice: "Can't save '#{group_params[:name]}'" }
+      format.json { render status: :error, json: "#{@group.name} not created" }
     end
   end
 
@@ -32,36 +37,40 @@ class GroupsController < ApplicationController
       @machine.groups.delete(@group)
     end
 
-    redirect_to group_path @group 
+    redirect_to group_path @group
   end
 
   def delete_user
-    @user = User.find(params[:user_id])
+    if current_user.admin? || @group.group_admin.user == current_user
+      @user = User.find(params[:user_id])
 
-    if @user.email.split('@').first != @group.name
-      @user.groups.each do |group|
-        REDIS_CACHE.del(GROUP_NAME_RESPONSE + group.name)
-        REDIS_CACHE.del(GROUP_GID_RESPONSE + group.gid.to_s)
+      if @user.email.split('@').first != @group.name
+        @user.groups.each do |group|
+          REDIS_CACHE.del(GROUP_NAME_RESPONSE + group.name)
+          REDIS_CACHE.del(GROUP_GID_RESPONSE + group.gid.to_s)
+        end
+        @user.groups.delete(@group)
+        REDIS_CACHE.del(PASSWD_NAME_RESPONSE + @user.email.split('@').first)
+        REDIS_CACHE.del(SHADOW_NAME_RESPONSE + @user.email.split('@').first)
+        REDIS_CACHE.del(PASSWD_UID_RESPONSE + @user.uid.to_s)
+
+        @response = Group.get_all_response.to_json
+        REDIS_CACHE.set(GROUP_ALL_RESPONSE, @response)
+        REDIS_CACHE.expire(GROUP_ALL_RESPONSE, REDIS_KEY_EXPIRY)
+        @response = User.get_all_shadow_response.to_json
+        REDIS_CACHE.set(SHADOW_ALL_RESPONSE, @response)
+        REDIS_CACHE.expire(SHADOW_ALL_RESPONSE, REDIS_KEY_EXPIRY)
+        @response = User.get_all_passwd_response.to_json
+        REDIS_CACHE.set(PASSWD_ALL_RESPONSE, @response)
+        REDIS_CACHE.expire(PASSWD_ALL_RESPONSE, REDIS_KEY_EXPIRY)
       end
-      @user.groups.delete(@group)
-      REDIS_CACHE.del(PASSWD_NAME_RESPONSE + @user.email.split('@').first)
-      REDIS_CACHE.del(SHADOW_NAME_RESPONSE + @user.email.split('@').first)
-      REDIS_CACHE.del(PASSWD_UID_RESPONSE + @user.uid.to_s)
 
-      @response = Group.get_all_response.to_json
-      REDIS_CACHE.set(GROUP_ALL_RESPONSE, @response)
-      REDIS_CACHE.expire(GROUP_ALL_RESPONSE, REDIS_KEY_EXPIRY)
-      @response = User.get_all_shadow_response.to_json
-      REDIS_CACHE.set(SHADOW_ALL_RESPONSE, @response)
-      REDIS_CACHE.expire(SHADOW_ALL_RESPONSE, REDIS_KEY_EXPIRY)
-      @response = User.get_all_passwd_response.to_json
-      REDIS_CACHE.set(PASSWD_ALL_RESPONSE, @response)
-      REDIS_CACHE.expire(PASSWD_ALL_RESPONSE, REDIS_KEY_EXPIRY)
+      VpnGroupUserAssociation.where(group_id: @group.id, user_id: params[:user_id]).destroy_all
     end
     redirect_to group_path @group
   end
 
-  def add_user 
+  def add_user
     if current_user.admin?
       user = User.find(params[:user_id])
       user.groups << @group if user.present? and user.groups.find_by_id(@group.id).blank?
@@ -138,11 +147,14 @@ class GroupsController < ApplicationController
   end
 
   def delete_vpn
-    VpnGroupAssociation.where(group_id: @group.id, vpn_id: params[:vpn_id]).destroy_all
+    if current_user.admin? || @group.group_admin.user == current_user
+      VpnGroupAssociation.where(group_id: @group.id, vpn_id: params[:vpn_id]).destroy_all
+      VpnGroupUserAssociation.where(group_id: @group.id, vpn_id: params[:vpn_id]).destroy_all
 
-    respond_to do |format|
-      format.html do
-        redirect_to group_path @group
+      respond_to do |format|
+        format.html do
+          redirect_to group_path @group
+        end
       end
     end
   end
